refactor(core): split service DTOs into types/results
