+++ src/local_rag_backend/app/services/__init__.py
@@ -1,5 +1,5 @@
 """Application-layer services."""
 
-from local_rag_backend.app.services import docs, index, ports
+from local_rag_backend.app.services import docs, index, ports, results
 
-__all__ = ["docs", "index", "ports"]
+__all__ = ["docs", "index", "ports", "results"]
